requirements.txt
flask
flask-cors
python-dotenv
flask-sqlalchemy
psycopg2-binary
faiss-cpu
huggingface-hub
requests
pypdf
python-docx
python-pptx
werkzeug
bcrypt
supabase
